Synced dispatch with core switch code. Spread dispatch jobs out more over a cluster. Fixed a bug in adding an infeasibility search to cplex_options. Try to make the ampl license manager exit after jobs are finished. Made dispatch input files group readable and writable so we can poke in each other's directories. Updated ssh tunnel connection to better handle user accounts having different names between machines. Made dispatch job management script more robust. Added code to monitor ampl processes as they are running. Retrieve cap factor inputs only for intermittent projects used in the main optimization (The strategies for handling and resuing those still need improvement). Fleshed out and expanded code to move results from AMPL into the database. Dispatch results are summarized by timepoint, load area, and technology. Load and balancing infeasibilities are exported along with the capacity expansions needed to address the infeasibilities.

+++ AMPL/dispatch/get_test_inputs.sh
@@ -24,20 +24,22 @@ function print_help {
 	head -n $last_line $0 | sed -e '/^#[ 	]/ !d' -e 's/^#[ 	]//'
 }
 
+# Set the umask to give group read & write permissions to all files & directories made by this script.
+umask=0022
 
 db_server="switch-db1.erg.berkeley.edu"
 DB_name="switch_inputs_wecc_v2_2"
 port=3306
 if [ $(hostname | grep 'citris' | wc -l) -gt 0 ]; then
-  base_data_dir="$HOME/shared/data/dispatch"
+  base_data_dir="$HOME/shared/data/dispatch/daily"
 else
-  base_data_dir="/Volumes/vtrak/switch_dispatch_all_weeks"
+  base_data_dir="/Volumes/switch/switch_dispatch_all_weeks/daily"
 fi
 if [ ! -d "$base_data_dir" ]; then 
   echo "Cannot find a base directory for the dispatch inputs at $base_data_dir."
   exit -1
 fi
-ssh_tunnel=0
+ssh_tunnel=1
 
 ###################################################
 # Detect optional command-line arguments
@@ -46,6 +48,8 @@ while [ -n "$1" ]; do
 case $1 in
   -t | --tunnel)
     ssh_tunnel=1; shift 1 ;;
+  --notunnel)
+    ssh_tunnel=0; shift 1 ;;
   -u)
     user=$2; shift 2 ;;
   -p)
@@ -115,7 +119,7 @@ if [ $ssh_tunnel -eq 1 ]; then
     local_port=$((local_port+1))
     is_port_free $local_port
   done
-  ssh -N -p 22 -c 3des $db_server -L $local_port/$db_server/$port &
+  ssh -N -p 22 -c 3des $user@$db_server -L $local_port/127.0.0.1/$port &
   ssh_pid=$!
   sleep 1
   connection_string="-h 127.0.0.1 --port $local_port -u $user -p$password $DB_name"
@@ -144,15 +148,33 @@ LOAD_SCENARIO_ID=$(mysql $connection_string --column-names=false -e "select load
 # Make links to the common input files in the parent directory instead of exporting from the DB again
 input_dir="common_inputs"
 mkdir -p $input_dir
-for f in balancing_areas.tab biomass_supply_curve_breakpoint.tab biomass_supply_curve_slope.tab carbon_cap_targets.tab existing_plants.tab fuel_costs.tab fuel_info.tab fuel_qualifies_for_rps.tab generator_info.tab load_areas.tab max_system_loads.tab misc_params.dat proposed_projects.tab rps_compliance_entity_targets.tab scenario_information.txt transmission_lines.tab; do
-  ln -sf ../../inputs/$f $input_dir/
+for f in balancing_areas.tab biomass_supply_curve_breakpoint.tab biomass_supply_curve_slope.tab carbon_cap_targets.tab existing_plants.tab fuel_costs.tab fuel_info.tab generator_costs.tab generator_info.tab load_areas.tab max_system_loads.tab misc_params.dat ng_supply_curve_slope.tab ng_supply_curve_breakpoint_consumption.tab ng_regional_price_adders.tab proposed_projects.tab rps_compliance_entity_targets.tab scenario_information.txt transmission_lines.tab ; do
+  if [ -f ../inputs/$f ]; then
+    ln -sf ../../inputs/$f $input_dir/
+  else
+    echo "Warning! $f does not exist in the input directory."
+  fi;
 done
+cd $input_dir;
 for b in InstallGen OperateEPDuringPeriod InstallTrans; do
-  for p in $(cd $input_dir; ls ../../results/$b*); do 
-    ln -sf $p $input_dir/
+  for p in $(ls ../../results/$b*); do 
+    ln -sf $p .
   done
 done
+cd ..
 ln -sf ../scenario_id.txt .
+
+####################################################
+# Goal: Make a list of project IDs used by the optimization that we need capacity factors for. 
+mkdir -p tmp
+projects_that_need_cap_factors=tmp/projects_that_need_cap_factors.txt
+# First make a list of technologies that need capacity factors. The criteria are taken from the AMPL check: check: card({(pid, a, t) in PROJECTS: intermittent[t] and resource_limited[t] and not ccs[t]} diff PROJ_INTERMITTENT) = 0;
+awk '{ if(NF >= 18) { tech=$1; intermittent=$17; resource_limited=$18; ccs=$16; if(intermittent == 1 && resource_limited == 1 && ccs == 0) { print tech; } } }' ../inputs/generator_info.tab > tmp/tech_that_needs_cap_factors.txt
+# Make a regular expression to match any member in the list of technologies
+techs_needed=$(awk '{if (NR>1 && length($1)>0) printf "|"$1; else if (NR==1 && length($1) > 0) printf $1; }' tmp/tech_that_needs_cap_factors.txt)
+# Use the tech list to pull project ids
+awk '{ if( $3 ~ /'$techs_needed'/) print $1; }' ../inputs/proposed_projects.tab > $projects_that_need_cap_factors
+
 
 ##########################
 # Make directories and gather inputs for each dispatch week in the study.
@@ -195,8 +217,6 @@ for test_set_id in $(mysql $connection_string --column-names=false -e "select di
 	# [rows of data]
 
 	SAMPLE_RESTRICTIONS="test_set_id=$test_set_id"
-	INTERMITTENT_PROJECTS_SELECTION="(( avg_cap_factor_percentile_by_intermittent_tech >= 0.75 or cumulative_avg_MW_tech_load_area <= 3 * total_yearly_load_mwh / 8766 or rank_by_tech_in_load_area <= 5 or avg_cap_factor_percentile_by_intermittent_tech is null) and technology <> 'Concentrating_PV')"
-
 	f="study_hours.tab"
 	echo "	$f..."
 	if [ ! -f $data_dir/$f ]; then
@@ -271,53 +291,60 @@ for test_set_id in $(mysql $connection_string --column-names=false -e "select di
 	echo "	$f..."
 	if [ ! -f $data_dir/$f ]; then
 		echo ampl.tab 4 1 > $data_dir/$f
+		echo "project_id  load_area technology  hour  cap_factor" >> $data_dir/$f
 		if [ $LOAD_SCENARIO_ID -lt 10 ]; then 
-      mysql $connection_string -e "\
-      select project_id, load_area, technology, DATE_FORMAT(datetime_utc,'%Y%m%d%H') as hour, cap_factor  \
-        FROM dispatch_test_sets \
-          JOIN study_timepoints USING(timepoint_id)\
-          JOIN _cap_factor_intermittent_sites ON(historic_hour=hour)\
-          JOIN _proposed_projects_v2 USING(project_id)\
-          JOIN load_area_info USING(area_id)\
-        WHERE training_set_id=$TRAINING_SET_ID \
-          AND $INTERMITTENT_PROJECTS_SELECTION\
-          AND test_set_id=$test_set_id;" >> $data_dir/$f
+		  while read project_id; do
+		    echo "\
+          select project_id, load_area, technology, DATE_FORMAT(datetime_utc,'%Y%m%d%H') as hour, cap_factor  \
+            FROM dispatch_test_sets \
+              JOIN study_timepoints USING(timepoint_id)\
+              JOIN _cap_factor_intermittent_sites ON(historic_hour=hour)\
+              JOIN _proposed_projects_v2 USING(project_id)\
+              JOIN load_area_info USING(area_id)\
+            WHERE training_set_id=$TRAINING_SET_ID \
+              AND project_id = $project_id\
+              AND test_set_id=$test_set_id;"
+		  done < $projects_that_need_cap_factors > tmp/select_cap_factors.sql 
+      mysql $connection_string -sN < tmp/select_cap_factors.sql >> $data_dir/$f
     else
-      mysql $connection_string -e "\
-      select project_id, load_area, technology, DATE_FORMAT(datetime_utc,'%Y%m%d%H') as hour, cap_factor  \
-        FROM dispatch_test_sets \
-          JOIN study_timepoints USING(timepoint_id)\
-          JOIN _cap_factor_intermittent_sites ON(historic_hour=hour)\
-          JOIN _proposed_projects_v2 USING(project_id)\
-          JOIN load_area_info USING(area_id)\
-        WHERE training_set_id=$TRAINING_SET_ID \
-          AND $INTERMITTENT_PROJECTS_SELECTION \
-          AND technology_id NOT IN (select technology_id from generator_info_v2 WHERE fuel='Solar')\
-          AND test_set_id=$test_set_id;" >> $data_dir/$f
-      # Get solar cap factors from 2005
-      mysql $connection_string -sN -e "\
+      # Get cap factors from 2006 for all non-solar projects
+		  while read project_id; do
+		    echo "\
+          select project_id, load_area, technology, DATE_FORMAT(datetime_utc,'%Y%m%d%H') as hour, cap_factor  \
+            FROM dispatch_test_sets \
+              JOIN study_timepoints USING(timepoint_id)\
+              JOIN _cap_factor_intermittent_sites ON(historic_hour=hour)\
+              JOIN _proposed_projects USING(project_id)\
+              JOIN load_area_info USING(area_id)\
+            WHERE training_set_id=$TRAINING_SET_ID \
+              AND project_id = $project_id \
+              AND technology_id NOT IN (select technology_id from generator_info WHERE fuel='Solar')\
+              AND test_set_id=$test_set_id;"
+      done < $projects_that_need_cap_factors > tmp/select_non_solar_cap_factors.sql 
+      mysql $connection_string -sN < tmp/select_non_solar_cap_factors.sql >> $data_dir/$f
+      # Get cap factors from 2005 for solar projects
+      echo "\
       CREATE TEMPORARY TABLE solar_tp_mapping \
         SELECT distinct historic_hour, historic_hour-8760 as hour_2005 from load_scenario_historic_timepoints WHERE load_scenario_id=$LOAD_SCENARIO_ID; \
-      ALTER TABLE solar_tp_mapping ADD INDEX ( historic_hour, hour_2005 ), ADD INDEX ( hour_2005, historic_hour ); \
-      select project_id, load_area, technology, DATE_FORMAT(datetime_utc,'%Y%m%d%H') as hour, cap_factor  \
-        FROM dispatch_test_sets \
-          JOIN study_timepoints USING(timepoint_id)\
-          JOIN solar_tp_mapping USING (historic_hour) \
-          JOIN _cap_factor_intermittent_sites ON(hour_2005=hour)\
-          JOIN _proposed_projects_v2 USING(project_id)\
-          JOIN load_area_info USING(area_id)\
-        WHERE training_set_id=$TRAINING_SET_ID \
-          AND test_set_id=$test_set_id \
-          AND $INTERMITTENT_PROJECTS_SELECTION \
-          AND technology_id IN (select technology_id from generator_info_v2 WHERE fuel='Solar');" >> $data_dir/$f
+      ALTER TABLE solar_tp_mapping ADD INDEX ( historic_hour, hour_2005 ), ADD INDEX ( hour_2005, historic_hour );" > tmp/select_solar_cap_factors.sql 
+		  while read project_id; do
+		    echo "\
+          select project_id, load_area, technology, DATE_FORMAT(datetime_utc,'%Y%m%d%H') as hour, cap_factor  \
+            FROM dispatch_test_sets \
+              JOIN study_timepoints USING(timepoint_id)\
+              JOIN solar_tp_mapping USING (historic_hour) \
+              JOIN _cap_factor_intermittent_sites ON(hour_2005=hour)\
+              JOIN _proposed_projects USING(project_id)\
+              JOIN load_area_info USING(area_id)\
+            WHERE training_set_id=$TRAINING_SET_ID \
+              AND test_set_id=$test_set_id \
+              AND project_id = $project_id \
+              AND technology_id IN (select technology_id from generator_info WHERE fuel='Solar');"
+      done < $projects_that_need_cap_factors >> tmp/select_solar_cap_factors.sql 
+      mysql $connection_string -sN < tmp/select_solar_cap_factors.sql >> $data_dir/$f
     fi
 	fi
 	[ -L $input_dir/$f ] && rm $input_dir/$f  # Remove the link if it exists
 	ln -s $data_dir/$f $input_dir/$f          # Make a new link
 
-# clean_up
-# exit
-
 done
-
-#clean_up
